{{cookiecutter.project_slug}}/requirements.txt
tartiflette-aiohttp
aiohttp-devtools==0.12